+++ src/grandorgue/model/GOEnclosure.cpp
@@ -1,6 +1,6 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2025 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2026 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
@@ -26,10 +26,24 @@ GOEnclosure::GOEnclosure(GOOrganModel &organModel)
     m_Displayed1(false),
     m_Displayed2(false),
     m_AmpMinimumLevel(0),
-    m_MIDIValue(0) {}
+    m_MIDIValue(0),
+    m_DefaultLowShelfFrequency(0.0f),
+    m_DefaultLowShelfAttenuationDb(0.0f),
+    m_DefaultHighShelfFrequency(0.0f),
+    m_DefaultHighShelfAttenuationDb(0.0f),
+    m_LowShelfFrequency(0.0f),
+    m_LowShelfAttenuationDb(0.0f),
+    m_HighShelfFrequency(0.0f),
+    m_HighShelfAttenuationDb(0.0f) {}
 
 static const wxString WX_AMP_MINIMUM_LEVEL = wxT("AmpMinimumLevel");
 static const wxString WX_VALUE = wxT("Value");
+static const wxString WX_LOW_SHELF_FREQUENCY = wxT("LowShelfFrequency");
+static const wxString WX_LOW_SHELF_ATTENUATION_DB
+  = wxT("LowShelfAttenuationDb");
+static const wxString WX_HIGH_SHELF_FREQUENCY = wxT("HighShelfFrequency");
+static const wxString WX_HIGH_SHELF_ATTENUATION_DB
+  = wxT("HighShelfAttenuationDb");
 
 void GOEnclosure::LoadFromCmb(GOConfigReader &cfg, uint8_t defaultValue) {
   m_AmpMinimumLevel = cfg.ReadInteger(
@@ -40,6 +54,38 @@ void GOEnclosure::LoadFromCmb(GOConfigReader &cfg, uint8_t defaultValue) {
     100,
     false,
     m_DefaultAmpMinimumLevel);
+  m_LowShelfFrequency = cfg.ReadFloat(
+    CMBSetting,
+    m_group,
+    WX_LOW_SHELF_FREQUENCY,
+    0,
+    0,
+    false,
+    m_DefaultLowShelfFrequency);
+  m_LowShelfAttenuationDb = cfg.ReadFloat(
+    CMBSetting,
+    m_group,
+    WX_LOW_SHELF_ATTENUATION_DB,
+    0,
+    121,
+    false,
+    m_DefaultLowShelfAttenuationDb);
+  m_HighShelfFrequency = cfg.ReadFloat(
+    CMBSetting,
+    m_group,
+    WX_HIGH_SHELF_FREQUENCY,
+    0,
+    0,
+    false,
+    m_DefaultHighShelfFrequency);
+  m_HighShelfAttenuationDb = cfg.ReadFloat(
+    CMBSetting,
+    m_group,
+    WX_HIGH_SHELF_ATTENUATION_DB,
+    0,
+    121,
+    false,
+    m_DefaultHighShelfAttenuationDb);
   SetEnclosureValue(cfg.ReadInteger(
     CMBSetting, m_group, WX_VALUE, 0, MAX_MIDI_VALUE, false, defaultValue));
 }
@@ -52,6 +98,10 @@ void GOEnclosure::Init(
   m_IsOdfDefined = false;
   GOMidiReceivingSendingObject::Init(cfg, group, name);
   m_DefaultAmpMinimumLevel = 0;
+  m_DefaultLowShelfFrequency = 0.0f;
+  m_DefaultLowShelfAttenuationDb = 0.0f;
+  m_DefaultHighShelfFrequency = 0.0f;
+  m_DefaultHighShelfAttenuationDb = 0.0f;
   LoadFromCmb(cfg, defaultValue);
 }
 
@@ -65,6 +115,16 @@ void GOEnclosure::Load(GOConfigReader &cfg, const wxString &group) {
     = cfg.ReadBoolean(ODFSetting, m_group, wxT("Displayed"), false, false);
   m_DefaultAmpMinimumLevel
     = cfg.ReadInteger(ODFSetting, m_group, WX_AMP_MINIMUM_LEVEL, 0, 100);
+  m_DefaultLowShelfAttenuationDb = cfg.ReadFloat(
+    ODFSetting, m_group, WX_LOW_SHELF_ATTENUATION_DB, 0, 121, false, 0);
+  m_DefaultLowShelfFrequency = m_DefaultLowShelfAttenuationDb > 0
+    ? cfg.ReadFloat(ODFSetting, m_group, WX_LOW_SHELF_FREQUENCY, 0, 0, true)
+    : 0.0f;
+  m_DefaultHighShelfAttenuationDb = cfg.ReadFloat(
+    ODFSetting, m_group, WX_HIGH_SHELF_ATTENUATION_DB, 0, 121, false, 0);
+  m_DefaultHighShelfFrequency = m_DefaultHighShelfAttenuationDb > 0
+    ? cfg.ReadFloat(ODFSetting, m_group, WX_HIGH_SHELF_FREQUENCY, 0, 0, true)
+    : 0.0f;
   LoadFromCmb(cfg, MAX_MIDI_VALUE);
 }
 
@@ -72,6 +132,11 @@ void GOEnclosure::Save(GOConfigWriter &cfg) {
   GOMidiReceivingSendingObject::Save(cfg);
   cfg.WriteInteger(m_group, WX_AMP_MINIMUM_LEVEL, m_AmpMinimumLevel);
   cfg.WriteInteger(m_group, WX_VALUE, m_MIDIValue);
+  cfg.WriteFloat(m_group, WX_LOW_SHELF_FREQUENCY, m_LowShelfFrequency);
+  cfg.WriteFloat(m_group, WX_LOW_SHELF_ATTENUATION_DB, m_LowShelfAttenuationDb);
+  cfg.WriteFloat(m_group, WX_HIGH_SHELF_FREQUENCY, m_HighShelfFrequency);
+  cfg.WriteFloat(
+    m_group, WX_HIGH_SHELF_ATTENUATION_DB, m_HighShelfAttenuationDb);
 }
 
 void GOEnclosure::SetEnclosureValue(uint8_t n) {
@@ -81,12 +146,6 @@ void GOEnclosure::SetEnclosureValue(uint8_t n) {
   }
   r_OrganModel.UpdateVolume();
   r_OrganModel.SendControlChanged(this);
-}
-
-float GOEnclosure::GetAttenuation() {
-  static const float scale = 1.0 / 12700.0;
-  return (float)(m_MIDIValue * (100 - m_AmpMinimumLevel) + 127 * m_AmpMinimumLevel)
-    * scale;
 }
 
 void GOEnclosure::Scroll(bool scroll_up) {