+++ src/tests/GOTestExe.cpp
@@ -17,6 +17,7 @@
 #include "testing/midi/GOTestMidiPlayerContent.h"
 #include "testing/midi/GOTestMidiSendProxy.h"
 #include "testing/model/GOTestDrawStop.h"
+#include "testing/model/GOTestEnclosure.h"
 #include "testing/model/GOTestOrganModel.h"
 #include "testing/model/GOTestPipeConfigTreeNode.h"
 #include "testing/model/GOTestSoundingPipe.h"
@@ -37,8 +38,11 @@
 #include "testing/sound/buffer/GOTestSoundBufferPlanar.h"
 #include "testing/sound/buffer/GOTestSoundBufferPlanarManaged.h"
 #include "testing/sound/buffer/GOTestSoundBufferPlanarMutable.h"
+#include "testing/sound/effects/GOTestPerfSoundShelfFilterProcessor.h"
 #include "testing/sound/effects/GOTestSoundReverbProcessor.h"
 #include "testing/sound/effects/GOTestSoundShelfFilterProcessor.h"
+#include "testing/sound/mappers/GOTestPerfSoundEnclosureShelfMapper.h"
+#include "testing/sound/mappers/GOTestSoundEnclosureShelfMapper.h"
 #include "testing/sound/playing/GOTestReleaseAlignTable.h"
 #include "testing/sound/playing/GOTestSoundStream.h"
 #include "testing/sound/playing/GOTestSoundToneBalanceFilter.h"
@@ -79,6 +83,7 @@ int main(int argc, char *argv[]) {
   GOTestOrganReader testOrganReader;
   GOTestOrganController testOrganController;
   GOTestDrawStop testDrawStop;
+  GOTestEnclosure testEnclosure;
   GOTestOrganModel testOrganModel;
   GOTestPipeConfigTreeNode testPipeConfigTreeNode;
   GOTestSoundingPipe testSoundingPipe;
@@ -105,6 +110,9 @@ int main(int argc, char *argv[]) {
   GOTestSoundReverb testSoundReverb;
   GOTestSoundReverbProcessor testSoundReverbProcessor;
   GOTestSoundShelfFilterProcessor testSoundShelfFilterProcessor;
+  GOTestPerfSoundShelfFilterProcessor testPerfSoundShelfFilterProcessor;
+  GOTestSoundEnclosureShelfMapper testSoundEnclosureShelfMapper;
+  GOTestPerfSoundEnclosureShelfMapper testPerfSoundEnclosureShelfMapper;
   GOTestSoundOrganEngine testSoundOrganEngine;
   GOTestSoundOrganEngineFactory testSoundOrganEngineFactory;
   GOTestSoundCallbackConnector testSoundCallbackConnector;